move erayzer helper functions
